fix(upgrade): 更新升级脚本

+++ scripts/upgrade.sh
@@ -20,7 +20,7 @@ fi
 
 if [ ! -d "$install_dir/jumpserver" ]; then
     if [ ! -d "$jumpserver_backup/jumpserver" ]; then
-        echo -e "\033[31m jumpserver 未安装或者目录不正确 \033[0m"
+        echo -e "\033[31m [ERROR] jumpserver 未安装或者目录不正确 \033[0m"
         exit 1
     fi
 fi
@@ -29,12 +29,27 @@ if [ ! -d "$PROJECT_DIR/$Upgrade_Version" ]; then
     mkdir -p $PROJECT_DIR/$Upgrade_Version
 fi
 
+if [ "${Version:0:1}" == "1" ] || [ "${Version:1:1}" -le "2" ] || [ "${Version:3:1}" -le "4" ]; then
+    echo -e "\033[33m* v2.5.0 开始, 数据库只支持 MySQL >= 5.7 *\033[0m"
+    if [ $DB_HOST == 127.0.0.1 ]; then
+        if [ ! "$(rpm -qa | grep mysql-community-server)" ]; then
+            echo -e "\033[31m [ERROR] mysql 未正确部署 \033[0m"
+            exit 1
+        fi
+    else
+        read -p "确定环境无误请按 y 继续升级: " a
+        if [ "$a" != "y" -a "$a" != "Y" ]; then
+            exit 1
+        fi
+    fi
+fi
+
 if [ ! -f "$PROJECT_DIR/$Upgrade_Version/jumpserver-$Upgrade_Version.tar.gz" ]; then
     timeout 60s wget -qO wget -qO $PROJECT_DIR/$Upgrade_Version/jumpserver-$Upgrade_Version.tar.gz https://github.com/jumpserver/jumpserver/releases/download/$Upgrade_Version/jumpserver-$Upgrade_Version.tar.gz || {
         rm -f $PROJECT_DIR/$Upgrade_Version/jumpserver-$Upgrade_Version.tar.gz
         wget -qO $PROJECT_DIR/$Upgrade_Version/jumpserver-$Upgrade_Version.tar.gz http://demo.jumpserver.org/download/jumpserver/$Upgrade_Version/jumpserver-$Upgrade_Version.tar.gz || {
             rm -f $PROJECT_DIR/$Upgrade_Version/jumpserver-$Upgrade_Version.tar.gz
-            echo -e "\033[31m 下载 jumpserver 失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
+            echo -e "\033[31m [ERROR] 下载 jumpserver 失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
             exit 1
         }
     }
@@ -46,7 +61,7 @@ if [ ! -f "$PROJECT_DIR/$Upgrade_Version/lina-$Upgrade_Version.tar.gz" ]; then
         rm -f $PROJECT_DIR/$Upgrade_Version/lina-$Upgrade_Version.tar.gz
         wget -qO $PROJECT_DIR/$Upgrade_Version/lina-$Upgrade_Version.tar.gz http://demo.jumpserver.org/download/lina/$Upgrade_Version/lina-$Upgrade_Version.tar.gz || {
             rm -f $PROJECT_DIR/$Upgrade_Version/lina-$Upgrade_Version.tar.gz
-            echo -e "\033[31m 下载 lina 失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
+            echo -e "\033[31m [ERROR] 下载 lina 失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
             exit 1
         }
   }
@@ -60,7 +75,7 @@ if [ ! -f "$PROJECT_DIR/$Upgrade_Version/luna-$Upgrade_Version.tar.gz" ]; then
         rm -f $PROJECT_DIR/$Upgrade_Version/luna-$Upgrade_Version.tar.gz
         wget -qO $PROJECT_DIR/$Upgrade_Version/luna-$Upgrade_Version.tar.gz http://demo.jumpserver.org/download/luna/$Upgrade_Version/luna-$Upgrade_Version.tar.gz || {
             rm -f $PROJECT_DIR/$Upgrade_Version/luna-$Upgrade_Version.tar.gz
-            echo -e "\033[31m 下载 luna 失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
+            echo -e "\033[31m [ERROR] 下载 luna 失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
             exit 1
         }
     }
@@ -78,14 +93,14 @@ fi
 
 if [ ! "$(docker images | grep jms_koko | grep $Upgrade_Version)" ]; then
     docker pull jumpserver/jms_koko:$Upgrade_Version || {
-        echo -e "\033[31m 下载 koko 镜像失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
+        echo -e "\033[31m [ERROR] 下载 koko 镜像失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
         exit 1
     }
 fi
 
 if [ ! "$(docker images | grep jms_guacamole | grep $Upgrade_Version)" ]; then
     docker pull jumpserver/jms_guacamole:$Upgrade_Version || {
-        echo -e "\033[31m 下载 guacamole 镜像失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
+        echo -e "\033[31m [ERROR] 下载 guacamole 镜像失败, 请检查网络是否正常或尝试重新执行升级脚本 \033[0m"
         exit 1
     }
 fi
@@ -117,7 +132,7 @@ fi
 source $install_dir/py3/bin/activate
 pip install --upgrade pip setuptools
 pip install -r $install_dir/jumpserver/requirements/requirements.txt || {
-    echo -e "\033[31m 升级 python 依赖失败, 请检查网络是否正常或者更换 pypi 源 \033[0m"
+    echo -e "\033[31m [ERROR] 升级 python 依赖失败, 请检查网络是否正常或者更换 pypi 源 \033[0m"
     exit 1
 }
 
@@ -130,7 +145,7 @@ if [ "${Version:0:1}" == "1" ]; then
     if [ ! -f "$PROJECT_DIR/$Upgrade_Version/jumpserver.conf" ]; then
         wget -qO $PROJECT_DIR/$Upgrade_Version/jumpserver.conf http://demo.jumpserver.org/download/nginx/conf.d/latest/jumpserver.conf || {
             rm -f $PROJECT_DIR/$Upgrade_Version/jumpserver.conf
-            echo "[ERROR] 下载 nginx 配置文件失败"
+            echo -e "\033[31m [ERROR] 下载 nginx 配置文件失败"
         }
     fi
     cp $PROJECT_DIR/$Upgrade_Version/jumpserver.conf /etc/nginx/conf.d/jumpserver.conf
@@ -152,12 +167,12 @@ if [ "${Version:0:1}" == "1" ]; then
 fi
 
 docker run --name jms_koko -d -p $ssh_port:2222 -p 127.0.0.1:5000:5000 -e CORE_HOST=http://$Server_IP:8080 -e BOOTSTRAP_TOKEN=$BOOTSTRAP_TOKEN --restart=always --privileged=true jumpserver/jms_koko:$Upgrade_Version || {
-    echo -e "\033[31m jms_koko 镜像下载失败, 请检查网络是否正常或者手动 pull 镜像 \033[0m"
+    echo -e "\033[31m [ERROR] jms_koko 镜像下载失败, 请检查网络是否正常或者手动 pull 镜像 \033[0m"
     exit 1
 }
 
 docker run --name jms_guacamole -d -p 127.0.0.1:8081:8080 -e JUMPSERVER_SERVER=http://$Server_IP:8080 -e BOOTSTRAP_TOKEN=$BOOTSTRAP_TOKEN --restart=always jumpserver/jms_guacamole:$Upgrade_Version || {
-    echo -e "\033[31m jms_guacamole 镜像下载失败, 请检查网络是否正常或者手动 pull 镜像 \033[0m"
+    echo -e "\033[31m [ERROR] jms_guacamole 镜像下载失败, 请检查网络是否正常或者手动 pull 镜像 \033[0m"
     exit 1
 }
